requirements.txt
scipy
librosa
tqdm
torchinfo
scipy
torch_geometric
tensorboard
pandas
